Updating organized code
