+++ Matrix.h
@@ -70,67 +70,3 @@ public:
         return out;
     }
 };
-
-/* Small matrix */
-/**** FIXME operator() const problem *****/
-template<typename Scalar, int rows, int cols, int StorageOrder = RowMajor>
-class SMatrix
-{
-public:
-    const int m_rows, m_cols;
-    Scalar m_data[rows*cols];
-
-    SMatrix() : m_rows(rows), m_cols(cols) {}
-
-    SMatrix(const Scalar& x0) : m_rows(rows), m_cols(cols)
-    {
-        setVal<rows, cols>(x0);
-    }
-
-    template<int r, int c>
-    void setVal(const Scalar& v)
-    {
-        for(int i = 0; i < r; i++)
-        {
-            for(int j = 0; j < c; j++)
-            {
-                m_data[StorageOrder == RowMajor ? i*cols + j : j*rows + i] = v;
-            }
-        }
-    }
-
-    inline Scalar& operator()(int i, int j) const
-    {
-        return m_data[StorageOrder == RowMajor ? i*cols + j : j*rows + i];
-    }
-
-    template<int mdepth, int mcols>
-    SMatrix<Scalar, rows, mdepth> mul(const SMatrix<Scalar, mdepth, mcols>& rhs)
-    {
-        SMatrix<Scalar, rows, mdepth> res;
-        for(int i = 0; i < rows; i++)
-        {
-            for(int j = 0; j < mcols; j++)
-            {
-                for(int k = 0; k < mdepth; k++)
-                {
-                    res.m_data[StorageOrder == RowMajor ? i*mdepth + j : j*rows + i] += m_data[StorageOrder == RowMajor ? i*mdepth + k : k*rows + i]*rhs.m_data[StorageOrder == RowMajor ? k*mcols + j : j*mdepth + k];
-                }
-            }
-        }
-        return res;
-    }
-
-    friend std::ostream& operator<<(std::ostream& out, const SMatrix& m)
-    {
-        for(int i = 0; i < rows;i++)
-        {
-            for(int j = 0; j < cols; j++)
-            {
-                out << m.m_data[StorageOrder == RowMajor ? i*cols + j : j*rows + i] << " ";
-            }
-            out << std::endl;
-        }
-        return out;
-    }
-};
