+++ gemm.cpp
@@ -1,5 +1,6 @@
 #include <iostream>
 #include "Matrix.h"
+#include "SMatrix.h"
 
 int main(int argc, char* argv[])
 {
